+++ InternetShop.Infrastructure/Persistence/Repository/UserRepository.cs
@@ -6,34 +6,37 @@ namespace InternetShop.Infrastructure.Persistence.Repository;
 public class UserRepository : IUserRepository
 {
     private readonly EfContext _dbContext;
-    public UserRepository(
+
+	public UserRepository(
         EfContext db)
     {
         _dbContext = db;
     }
-    public async Task<List<Order>> GetOrderListAsync(Guid userId, CancellationToken cancellationToken)
+    public async Task<IEnumerable<Order>> GetOrderListAsync(Guid userId, CancellationToken cancellationToken)
     {
-        if (userId == null)
+        if (userId == Guid.Empty)
         {
-            throw new Exception("Запрос пустой.");
-        }
-        var orders = _dbContext.Orders
+			throw new ArgumentException("Идентификатор пользователя не может быть пустым.", nameof(userId));
+		}
+        var orders = await _dbContext.Orders
             .Include(u => u.User)
-            .Where(o => o.User.Id == userId).ToList();
+            .Where(o => o.User.Id == userId).ToListAsync();
 
         if (orders == null || orders.Count == 0)
         {
-            throw new Exception("У пользователя с данным идентификатором отсутствуют заказы");
-        }
+			throw new KeyNotFoundException($"Заказы для пользователя с идентификатором {userId} не найдены.");
+		}
 
         return orders;
     }
 
-    public DbSet<User> Entities => throw new NotImplementedException();
-
     public async Task AddAsync(User entity)
-    {
-        await _dbContext.Users.AddAsync(entity);
+	{
+		if (entity == null)
+		{
+			throw new ArgumentException("Добавляемая сущность не может быть пустой.", nameof(entity));
+		}
+		await _dbContext.Users.AddAsync(entity);
         await _dbContext.SaveChangesAsync();
     }
 
@@ -42,12 +45,21 @@ public class UserRepository : IUserRepository
         var result = await _dbContext.Users
             .ToListAsync();
 
-        return result;
+		if (result == null)
+		{
+			throw new ArgumentException("Список пользователей в бд пуст.", nameof(result));
+		}
+
+		return result;
     }
 
     public async Task<User?> GetByIdAsync(Guid id)
-    {
-        var result = await _dbContext.Users
+	{
+		if (id == Guid.Empty)
+		{
+			throw new ArgumentException("Идентификатор пользователя не может быть пустым.", nameof(id));
+		}
+		var result = await _dbContext.Users
             .AsNoTracking()
             .FirstOrDefaultAsync(u => u.Id == id);
 
@@ -55,15 +67,25 @@ public class UserRepository : IUserRepository
     }
 
     public async Task RemoveAsync(User entity)
-    {
-        _dbContext.Users.Remove(entity);
+	{
+		if (entity == null)
+		{
+			throw new ArgumentException("Удаляемая сущность не может быть пустой.", nameof(entity));
+		}
+
+		_dbContext.Users.Remove(entity);
 
         await _dbContext.SaveChangesAsync();
     }
 
     public async Task UpdateAsync(User entity)
-    {
-        _dbContext.Users.Update(entity);
+	{
+		if (entity == null)
+		{
+			throw new ArgumentException("Обновляемая сущность не может быть пустой.", nameof(entity));
+		}
+
+		_dbContext.Users.Update(entity);
 
         await _dbContext.SaveChangesAsync();
     }
